Add blog/feed.xml
This is the atom version of the feed.  Largely this is taken from the old ubxd_web app/views/posts/feed.atom.builder.  We may have a problem with the id's for the site and articles being different (it's possible that we end up re-publishing everything because the ids are different).

We need redirects in place for the old urls to the new one.