Use hex colour directly instead of declaring them in variables file
Taking a different route to how colour is used across the stylesheets.
As we have well described and maintained partials, it is easy to find
colours scoped to blocks (seeing as we are following BEM). This
variables file is becoming difficult to maintain and is not giving us
any benefits.