Make Carrie and Willa images bigger on people page
Don’t want upload a different set of images as 
they are used in other places on the site which
would require styling changes anyway. 

These 2 caricatures look smaller than the others
because the poses they are in make them so wide,
and their dimensions are the same size. 

You can’t notice that on the people page because 
their bodies are hidden. I think adding a styling
exception for them is the least messy solution.