Enable posts to have a list of authors
Sometimes posts have more than one author, it'd be nice to represent
this on post pages.

This reads from `author` or `authors` from the front matter to build a
list of authors. The list of authors is used to draw the page (list of
names at the top, list of 'profiles' at the bottom).