Setup rspec correctly
We have a few specs in the project, but no .rspec or spec_helper.rb file to
configure it.  This commit adds one.  We also require files by name, rather
than direct path references into ./lib.  Finally, we rename article_sorter_spec
to related_blog_articles_spec, as RelatedBlogArticles is the name of the
object under test, not ArticleSorter (which doesn't exist).