Add a Tiler class
This class is to replace packery by doing the tiling at html render time,
rather than in JS on the browser.  We construct a Tiler with a number of
rows and columns.  When we add a tile we say what size it is (default is
1x1) and when we layout the tiles the tiler will fit things in and fill
up spaces created by adding larger tiles using any future smaller
tiles if possible.