Fix the spec for articles_by_author
Lulled into a false sense of security about why the build was failing (
I thought it was because rspec wasn't properly configured which I fixed
in a separate PR) I merged this without properly checking the spec results.

The problem was that we were trying to stub `blog` but also creating a
let variable called `blog`.  Changing the name of the let variable exposed
a different problem: the Helpers module doesn't implement `blog`
directly (it comes from middleman-blog) and so we need to provide a default
implementation instead of using a stub because we've configured rspec to
not let you stub methods that haven't been implemented.