Create a script to remove leavers
Removing leavers involves removing their section in the
`data/people.yml` as well as clearing out any links to the unboxed
`/people#name-of-person`. This can be fairly longwinded and it's easy to
miss links by hand.
This commit aims to make that easier by adding a rake task to remove a
person from the unboxed site. This is done by running `rake
remove_person` from the root directory. You can then type in the full
name of the person to remove (as written in the `people.yml` - i.e.
Abraao Mota, not abraao-mota), and the script does the rest.