Give reasonable default srcset value for external images
This helper method is used throughout the site to reference locally
hosted retina images according to our '@2x' convention. Components shown in the
style guide make use of placeholder images hosted externally but they were not
showing due to a problem with the srcset value. This commit fixes that
problem by changing the helper method so that passing in an absolute image URL
will return that same URL.