Fix object-fit styling on IE and Edge
The CSS property object-fit is not supported in older IE browsers and
Edge (though it's currently being implemented for Edge) resulting in
images being skewd when window is resized. To fix this we include a
javascript file created by bfred-it on github (link:
https://github.com/bfred-it/object-fit-images/).

As explained in his repo, by including the styling "font-family:
'object-fit: cover' " in the stylings for the image containers, we can
fix this problem.