Fix safari bug with blog tiles
There is a bug in Safari when a user first loads the blog page that
results in large gaps inbetween tiles. Subsequent loads of the page is
fine though. Investigation has led me to believe that the blog tiles
are arranged by a javascript plugin `Packery` in blog_grid.erb.
I suspect that the script is loading before the page is fully ready
resulting in some sort of calculation error in the grid arrangement.
By adding a window.onload, we can be sure that the script only fires
when the DOM is ready.