Use h2 not h1 in euruko blog post
The blog post title is the h1 and we shouldn't really have more than one per page.  We also can't tell what is a link and what isn't with an h1 as the whole think is the purple link colour - with an h2 only the link is purple, the rest of the text is classic black.