Changing titles of some old roundups to new style
We think that the old style titles end up looking like auto-generated
posts that mean people coming to the index page don't click on them.
We're trying giving them a different style of title which is a summary
of the topics included in the links to see if this helps.