Fix custom author indexes
In middleman 4 we can't directly access data in config.rb as it's not
quite initialized yet.  Instead of an array we get a LazyCollectorStep.
In this case that means we don't detect any authors (as everyone ends up
as a NonEmployee) and so we don't generate the `/blog/author/person/`
pages for them.  If we switch to using `@app.data` instead then we get
the correct version of the data (e.g. an Array, not a lazy object) and
so our filter works again.