Regenerate people.yml from new script that modifies file
When running the `lib/remove_person.rb` script, the people.yml file is
rewritten using the YAML library. Thankfully, order is preserved when
this happens, but we do lose some of the formatting that we've added in
(quotation marks, new lines, etc.) However, should should not be
problematic if we continue to use scripts to modify this file.