Specify Netfliy CMS config file to allow load from /admin and /admin/
To boot the Netlify CMS app, the URL that we visit must be /admin/.
However, the first time using the app, it was not obvious that the
trailing forward slash was required, and I dug into why the route didn't
work (thinking it was an issue with my config.yml - validation, syntax,
etc.)

However, we can get around this confusion by providing the config file
as a 'cms-config-url' link type in the admin/index.html page. Now we can
use /admin or /admin/, both of which will boot the Netlify CMS app.