Convert to circleci v2 config
We started by using the [config-translation endpoint][config-translation] that
circleci provides to convert our existing v1 config into the new v2 config.  It
contained lots of extra stuff and comments so I removed a load of them and
simplified it.

Note that we need to run our specs on a ruby image that matches the version
specified in our `.ruby-version`, and that we also need a ruby image that has
a JS runtime, so we use `circleci/ruby:2.3.7-node`.

[config-translation]: https://circleci.com/docs/2.0/config-translation/