Fix some more insecure content
All these changes are to convert http urls in `<img src="...">` and turn them
into https versions.  All the conversions are by extracting the image from the
bit.ly link (which is actually a photobucked link) and uploading them to our S3
bucket (unboxed-web-image-uploader) and using that url instead.  To make sure
the image names are unique we turn them all into uuids, which isn't very useful
in terms of knowing what the image is by filename, but at least we don't
overwrite any existing images.